scm: size_t expected!
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@45242 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/bldprogs/scm.cpp
@@ -449,7 +449,7 @@ static int scmSettingsBaseHandleOpt(PSCMSETTINGSBASE pSettings, int rc, PRTGETOP
                 return VINF_SUCCESS;
 
             return RTStrAAppendExN(ppsz, 2,
-                                   "|", *ppsz && **ppsz ? 1 : 0,
+                                   "|", *ppsz && **ppsz ? (size_t)1 : (size_t)0,
                                    pszSrc, cchSrc);
         }
